schema parsing CHANGE try to differentiate XPath errors
We are checking XPath expressions (of must and when) already when parsing
the schema. If the mentioned nodes are not found in the schemas, the
expression will always fail. Now we try to differentiate situation, when
it is certain error (the expression refers the node in the same schema
which is missing) and when it is possibly caused by not implementing another
module (the expresison refers the node from another module and this node is
not present). In the second case, we not report the issue as error, but only
as a warning. The schema is then correctly parsed, but the data cannot
appear anyway, since the must/when expression will always fail.

+++ src/common.h
@@ -238,7 +238,14 @@ enum LY_VLOG_ELEM {
     LY_VLOG_LYD, /* struct lyd_node* */
     LY_VLOG_STR  /* const char* */
 };
-void ly_vlog_hide(int hide);
+
+/*
+ * 0 - normal visibility
+ * 1-254 - do not print messages
+ * 255 - convert errors to warnings
+ */
+void ly_vlog_hide(uint8_t hide);
+
 uint8_t *ly_vlog_hide_location(void);
 void ly_vlog(LY_ECODE code, enum LY_VLOG_ELEM elem_type, const void *elem, ...);
 #define LOGVAL(code, elem_type, elem, args...)                      \
